# alvobot-reference/specs/019-alvoads-meta/contracts/api.yaml
openapi: 3.0.3
info:
  title: AlvoADS Meta API
  description: API endpoints for Meta/Facebook Ads campaign management
  version: 1.0.0
servers:
  - url: /api/meta/campaigns
    description: Meta Campaigns API
paths:
...
  /{id}/pause:
    post:
      summary: Pause published campaign
      operationId: pauseCampaign
      tags:
        - Publishing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign paused
        '404':
          $ref: '#/components/responses/NotFound'
...
  /targeting/countries:
    get:
      summary: Search countries for targeting
      operationId: searchCountries
      tags:
        - Targeting
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of countries
          content:
            application/json:
              schema:
                type: object
                properties:
                  countries:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeoTarget'
  /targeting/languages:
    get:
      summary: Search languages for targeting
      operationId: searchLanguages
      tags:
        - Targeting
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of languages
          content:
            application/json:
              schema:
                type: object
                properties:
                  languages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Language'

  # ============================================
  # AI Generation
  # ============================================
...
components:
  schemas:
...
    GeoTarget:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        type:
          type: string
        countryCode:
          type: string
    Language:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...